test gzip compression for static routes
