+++ Sources/DestinyMacros/extensions/HTTPResponseMessageExtensions.swift
@@ -3,7 +3,10 @@ import Destiny
 
 // MARK: HTTPResponseMessage
 extension HTTPResponseMessage {
-    public func intermediateString(escapeLineBreak: Bool) -> String {
+    public func headString(
+        escapeLineBreak: Bool,
+        contentLength: Int
+    ) -> String {
         let suffix = escapeLineBreak ? "\\r\\n" : "\r\n"
         var string = head.string(suffix: suffix)
         if let body {
@@ -11,7 +14,7 @@ extension HTTPResponseMessage {
                 string += "content-type: \(contentType)\((charset != nil ? "; charset=\(charset!.rawName)" : ""))\(suffix)"
             }
             if body.hasContentLength {
-                string += "content-length: \(body.count)\(suffix)\(suffix)"
+                string += "content-length: \(contentLength)\(suffix)\(suffix)"
             }
         }
         return string
