+++ Sources/DestinyMacros/util/IntermediateResponseBody.swift
@@ -8,8 +8,13 @@ public struct IntermediateResponseBody: ResponseBodyProtocol {
     public let valueExpr:ExprSyntax
     public var type:IntermediateResponseBodyType
     let value:String
-    public let count:Int
-    private var interpolation = 0
+    public private(set) var count:Int
+    var interpolation = 0
+    var rawValue:[UInt8]? = nil {
+        didSet {
+            count = rawValue?.count ?? count
+        }
+    }
 
     public init(
         type: IntermediateResponseBodyType,
@@ -29,18 +34,20 @@ public struct IntermediateResponseBody: ResponseBodyProtocol {
         self.value = valueString
         self.count = valueString.count - count
     }
-    private init(
+    init(
         valueExpr: ExprSyntax,
         type: IntermediateResponseBodyType,
         value: String,
         count: Int,
-        interpolation: Int
+        interpolation: Int,
+        rawValue: [UInt8]? = nil
     ) {
         self.valueExpr = valueExpr
         self.type = type
         self.value = value
-        self.count = count
+        self.count = rawValue?.count ?? count
         self.interpolation = interpolation
+        self.rawValue = rawValue
     }
 
     private static func upgradeSegments(_ list: StringLiteralSegmentListSyntax) -> (String, Int) {
@@ -89,12 +96,6 @@ public struct IntermediateResponseBody: ResponseBodyProtocol {
     }
 
     public func write(to buffer: UnsafeMutableBufferPointer<UInt8>, at index: inout Int) {
-    }
-
-    private func preDateAndPostDateValues(_ string: String) -> (preDate: Substring, postDate: Substring) {
-        let preDate = string[string.startIndex..<string.index(string.startIndex, offsetBy: 22)]
-        let postDate = string[string.index(string.startIndex, offsetBy: 51)...]
-        return (preDate, postDate)
     }
 
     var isNoncopyable: Bool {
@@ -169,252 +170,6 @@ extension IntermediateResponseBody {
         }
         context.diagnose(DiagnosticMsg.unhandled(node: expr))
         return nil
-    }
-}
-
-// MARK: Responder debug description
-extension IntermediateResponseBody {
-    func responderDebugDescription(
-        context: some MacroExpansionContext,
-        isCopyable: Bool,
-        responseString: inout String
-    ) -> String {
-        let prefix = isCopyable ? "" : "NonCopyable"
-        switch type {
-        case .bytes:
-            return "\(prefix)Bytes(\(bytesPayload(context: context, responseString: &responseString)))"
-        case .inlineBytes:
-            return "\(prefix)InlineBytes(\(bytesPayload(context: context, responseString: &responseString)))"
-        case .macroExpansion:
-            responseString.removeLast(8 + String(value.count).count) // "#\r\n\r\n".count
-            return "RouteResponses.\(prefix)MacroExpansion(\"\(responseString)\", body: \(value))"
-        case .macroExpansionWithDateHeader:
-            var (preDate, postDate) = preDateAndPostDateValues(responseString)
-            postDate.removeLast(8 + String(value.count).count) // "#\r\n\r\n".count
-            return "\(prefix)MacroExpansionWithDateHeader(preDateValue: \"\(preDate)\", postDateValue: \"\(postDate)\", body: \(value))"
-        case .streamWithDateHeader:
-            var (preDate, postDate) = preDateAndPostDateValues(responseString)
-            postDate = "\\r\\nTransfer-Encoding: chunked\(postDate)"
-            return "\(prefix)StreamWithDateHeader(preDateValue: \"\(preDate)\", postDateValue: \"\(postDate)\\r\\n\", body: \(value))"
-        case .stringWithDateHeader:
-            if interpolation == 0 {
-                // upgrade
-                return IntermediateResponseBody(
-                    valueExpr: valueExpr,
-                    type: .staticStringWithDateHeader,
-                    value: escapedValue(),
-                    count: count,
-                    interpolation: interpolation
-                ).responderDebugDescription(context: context, isCopyable: isCopyable, responseString: &responseString)
-            }
-            let delimiter = valueExpr.stringLiteral?.openingPounds?.text ?? ""
-            let (preDate, postDate) = preDateAndPostDateValues(responseString)
-            return "StringWithDateHeader(preDateValue: \(delimiter)\"\(preDate)\"\(delimiter), postDateValue: \(delimiter)\"\(postDate)\"\(delimiter), value: \(delimiter)\"\(escapedValue())\"\(delimiter))"
-        case .staticString:
-            let delimiter = valueExpr.stringLiteral?.openingPounds?.text ?? ""
-            return "StaticString(\(delimiter)\"\(responseString)\(escapedValue())\"\(delimiter))"
-        case .staticStringWithDateHeader:
-            let delimiter = valueExpr.stringLiteral?.openingPounds?.text ?? ""
-            let (preDate, postDate) = preDateAndPostDateValues("\(responseString)\(escapedValue())")
-            return "\(prefix)StaticStringWithDateHeader(preDateValue: \(delimiter)\"\(preDate)\"\(delimiter), postDateValue: \(delimiter)\"\(postDate)\"\(delimiter))"
-
-        case .string:
-            var s = responseString + value
-            if s.first != "\"" {
-                s.insert("\"", at: s.startIndex)
-            }
-            if s.last != "\"" {
-                s.append("\"")
-            }
-            if let stringLiteral = valueExpr.stringLiteral, let openingPounds = stringLiteral.openingPounds, let closingPounds = stringLiteral.closingPounds {
-                s = openingPounds.text + s + closingPounds.text
-            }
-            return s
-
-        case .nonCopyableBytes:
-            return "NonCopyableBytes(\(bytesPayload(context: context, responseString: &responseString)))"
-        case .nonCopyableInlineBytes:
-            return "NonCopyableInlineBytes(\(bytesPayload(context: context, responseString: &responseString)))"
-        case .nonCopyableMacroExpansionWithDateHeader:
-            var (preDate, postDate) = preDateAndPostDateValues("\(responseString)")
-            postDate.removeLast(8 + String(value.count).count) // "#\r\n\r\n".count
-            return "NonCopyableMacroExpansionWithDateHeader(preDateValue: \"\(preDate)\", postDateValue: \"\(postDate)\", body: \(value))"
-        case .nonCopyableStreamWithDateHeader:
-            var (preDate, postDate) = preDateAndPostDateValues(responseString)
-            postDate = "\\r\\nTransfer-Encoding: chunked\(postDate)"
-            return "NonCopyableStreamWithDateHeader(preDateValue: \"\(preDate)\", postDateValue: \"\(postDate)\\r\\n\", body: \(value))"
-        case .nonCopyableStaticStringWithDateHeader:
-            let delimiter = valueExpr.stringLiteral?.openingPounds?.text ?? ""
-            let (preDate, postDate) = preDateAndPostDateValues("\(responseString)\(escapedValue())")
-            return "NonCopyableStaticStringWithDateHeader(preDateValue: \(delimiter)\"\(preDate)\"\(delimiter), postDateValue: \(delimiter)\"\(postDate)\"\(delimiter))"
-        }
-    }
-    func escapedValue() -> String {
-        var string = value
-        guard valueExpr.stringLiteral?.openingPounds == nil else {
-            // don't escape the string if it uses pound delimiters
-            return string
-        }
-        string.replace("\"", with: "\\\"")
-        return string
-    }
-
-    private func bytesPayload(
-        context: some MacroExpansionContext,
-        responseString: inout String
-    ) -> [UInt8] {
-        var payload = [UInt8]()
-        payload.reserveCapacity(responseString.count)
-        responseString.withUTF8 {
-            payload.append(contentsOf: $0)
-        }
-        if let elements = valueExpr.array?.elements {
-            for element in elements {
-                if let s = element.expression.integerLiteral?.literal.text, let byte = UInt8(s) {
-                    payload.append(byte)
-                } else if let s = element.expression.memberAccess?.declName.baseName.text, let byte = UInt8(convenientName: s) {
-                    payload.append(byte)
-                } else {
-                    context.diagnose(DiagnosticMsg.unhandled(node: element.expression))
-                }
-            }
-        } else {
-            context.diagnose(DiagnosticMsg.unhandled(node: valueExpr))
-        }
-        return payload
-    }
-}
-
-extension IntermediateResponseBody {
-    #if hasFeature(Embedded) || EMBEDDED
-    public func responderDebugDescription<B>(
-        context: some MacroExpansionContext,
-        isCopyable: Bool,
-        response: HTTPResponseMessage<B>
-    ) -> String {
-        let escapeLineBreak = !(type == .bytes || type == .nonCopyableBytes || type == .inlineBytes || type == .nonCopyableInlineBytes)
-        var responseString = response.intermediateString(escapeLineBreak: escapeLineBreak)
-        return responderDebugDescription(context: context, isCopyable: isCopyable, responseString: &responseString)
-    }
-    #else
-    public func responderDebugDescription(
-        context: some MacroExpansionContext,
-        isCopyable: Bool,
-        response: HTTPResponseMessage
-    ) -> String {
-        let escapeLineBreak = !(type == .bytes || type == .nonCopyableBytes || type == .inlineBytes || type == .nonCopyableInlineBytes)
-        var responseString = response.intermediateString(escapeLineBreak: escapeLineBreak)
-        return responderDebugDescription(context: context, isCopyable: isCopyable, responseString: &responseString)
-    }
-    #endif
-}
-
-// MARK: IntermediateResponseBodyType
-public enum IntermediateResponseBodyType: String, Sendable {
-    case bytes
-    case inlineBytes                             = "inlinebytes"
-    case macroExpansion                          = "macroexpansion"
-    case macroExpansionWithDateHeader            = "macroexpansionwithdateheader"
-    case streamWithDateHeader                    = "streamwithdateheader"
-    case staticString                            = "staticstring"
-    case staticStringWithDateHeader              = "staticstringwithdateheader"
-    case stringWithDateHeader                    = "stringwithdateheader"
-
-    case string
-
-    case nonCopyableBytes                        = "noncopyablebytes"
-    case nonCopyableInlineBytes                  = "noncopyableinlinebytes"
-    case nonCopyableMacroExpansionWithDateHeader = "noncopyablemacroexpansionwithdateheader"
-    case nonCopyableStreamWithDateHeader         = "noncopyablestreamwithdateheader"
-    case nonCopyableStaticStringWithDateHeader   = "noncopyablestaticstringwithdateheader"
-}
-
-extension IntermediateResponseBodyType {
-    public var isEnabled: Bool? {
-        switch self {
-        case .bytes:
-            #if CopyableBytes
-            true
-            #else
-            false
-            #endif
-        case .inlineBytes:
-            #if CopyableInlineBytes
-            true
-            #else
-            false
-            #endif
-        case .macroExpansion:
-            #if CopyableMacroExpansion
-            true
-            #else
-            false
-            #endif
-        case .macroExpansionWithDateHeader:
-            #if CopyableMacroExpansionWithDateHeader
-            true
-            #else
-            false
-            #endif
-        case .streamWithDateHeader:
-            #if CopyableStreamWithDateHeader
-            true
-            #else
-            false
-            #endif
-        case .staticString:
-            true
-        case .staticStringWithDateHeader:
-            #if CopyableStaticStringWithDateHeader
-            true
-            #else
-            false
-            #endif
-        case .stringWithDateHeader:
-            #if CopyableStringWithDateHeader
-            true
-            #else
-            false
-            #endif
-
-        case .string:
-            #if StringRouteResponder
-            true
-            #else
-            false
-            #endif
-
-        case .nonCopyableBytes:
-            #if NonCopyableBytes
-            true
-            #else
-            false
-            #endif
-        case .nonCopyableInlineBytes:
-            #if NonCopyableInlineBytes
-            true
-            #else
-            false
-            #endif
-        case .nonCopyableMacroExpansionWithDateHeader:
-            #if NonCopyableMacroExpansionWithDateHeader
-            true
-            #else
-            false
-            #endif
-        case .nonCopyableStreamWithDateHeader:
-            #if NonCopyableStreamWithDateHeader
-            true
-            #else
-            false
-            #endif
-        case .nonCopyableStaticStringWithDateHeader:
-            #if NonCopyableStaticStringWithDateHeader
-            true
-            #else
-            false
-            #endif
-        }
     }
 }
