+++ Sources/DestinyMacros/util/StaticRoute.swift
@@ -9,7 +9,7 @@ import SwiftSyntaxMacros
 public struct StaticRoute: Sendable {
     public var path:[String]
     public let contentType:String?
-    public let body:IntermediateResponseBody?
+    public internal(set) var  body:IntermediateResponseBody?
 
     public var method:HTTPRequestMethod
     public let status:HTTPResponseStatus.Code
@@ -71,120 +71,10 @@ extension StaticRoute {
     }
 }
 
-// MARK: Response
-extension StaticRoute {
-    #if StaticMiddleware
-        public func response(
-            middleware: [StaticMiddleware]
-        ) -> HTTPResponseMessage {
-            var version = version
-            let path = path.joined(separator: "/")
-            var status = status
-            var contentType = contentType
-            var headers = HTTPHeaders()
-            if body?.hasDateHeader ?? false {
-                headers["date"] = HTTPDateFormat.placeholder
-            }
-
-            #if HTTPCookie
-            var cookies = [HTTPCookie]()
-            #endif
-
-            middleware.forEach { middleware in
-                if middleware.handles(version: version, path: path, method: method, contentType: contentType, status: status) {
-                    #if HTTPCookie
-                    middleware.apply(version: &version, contentType: &contentType, status: &status, headers: &headers, cookies: &cookies)
-                    #else
-                    middleware.apply(version: &version, contentType: &contentType, status: &status, headers: &headers)
-                    #endif
-                }
-            }
-            headers["content-type"] = nil
-            headers["content-length"] = nil
-
-            #if HTTPCookie
-            return Self.response(
-                version: version,
-                status: status,
-                headers: &headers,
-                cookies: cookies,
-                body: body,
-                contentType: contentType,
-                charset: charset
-            )
-            #else
-            return Self.response(
-                version: version,
-                status: status,
-                headers: &headers,
-                body: body,
-                contentType: contentType,
-                charset: charset
-            )
-            #endif
-        }
-    #else
-        public func response() -> HTTPResponseMessage {
-            var headers = HTTPHeaders()
-            if body?.hasDateHeader ?? false {
-                headers["date"] = HTTPDateFormat.placeholder
-            }
-            headers["content-type"] = nil
-            headers["content-length"] = nil
-            #if HTTPCooke
-            return Self.response(version: version, status: status, headers: &headers, cookies: [], body: body, contentType: contentType, charset: charset)
-            #else
-            return Self.response(version: version, status: status, headers: &headers, body: body, contentType: contentType, charset: charset)
-            #endif
-        }
-    #endif
-
-    #if HTTPCookie
-    @inline(__always)
-    package static func response(
-        version: HTTPVersion,
-        status: HTTPResponseStatus.Code,
-        headers: inout HTTPHeaders,
-        cookies: [HTTPCookie],
-        body: IntermediateResponseBody?,
-        contentType: String?,
-        charset: Charset?
-    ) -> HTTPResponseMessage {
-        headers["content-type"] = nil
-        headers["content-length"] = nil
-        return HTTPResponseMessage(
-            head: .init(headers: headers, cookies: cookies, status: status, version: version),
-            body: body,
-            contentType: contentType,
-            charset: charset
-        )
-    }
-    #else
-    @inline(__always)
-    package static func response(
-        version: HTTPVersion,
-        status: HTTPResponseStatus.Code,
-        headers: inout HTTPHeaders,
-        body: IntermediateResponseBody?,
-        contentType: String?,
-        charset: Charset?
-    ) -> HTTPResponseMessage {
-        headers["content-type"] = nil
-        headers["content-length"] = nil
-        return HTTPResponseMessage(
-            head: .init(headers: headers, status: status, version: version),
-            body: body,
-            contentType: contentType,
-            charset: charset
-        )
-    }
-    #endif
-}
-
 // MARK: Responder
 extension StaticRoute {
     #if StaticMiddleware
-    public func responder(
+    public mutating func responder(
         middleware: [StaticMiddleware]
     ) -> String? {
         return response(middleware: middleware).string(escapeLineBreak: true)
@@ -194,44 +84,7 @@ extension StaticRoute {
         return response().string(escapeLineBreak: true)
     }
     #endif
-}
 
-// MARK: Response
-extension StaticRoute {
-    /// Builds the HTTP Message for this route.
-    /// 
-    /// - Parameters:
-    ///   - context: Macro expansion context where it was called.
-    ///   - function: `FunctionCallExprSyntax` that represents this route.
-    ///   - middleware: Static middleware this route will handle.
-    #if StaticMiddleware
-    public func response(
-        context: some MacroExpansionContext,
-        function: FunctionCallExprSyntax,
-        middleware: [StaticMiddleware]
-    ) -> HTTPResponseMessage {
-        let result = response(middleware: middleware)
-        if result.statusCode() == 501 { // not implemented
-            Diagnostic.routeResponseStatusNotImplemented(context: context, node: function.calledExpression)
-        }
-        return result
-    }
-    #else
-    public func response(
-        context: some MacroExpansionContext,
-        function: FunctionCallExprSyntax
-    ) -> HTTPResponseMessage {
-        let result = response()
-        if result.statusCode() == 501 { // not implemented
-            Diagnostic.routeResponseStatusNotImplemented(context: context, node: function.calledExpression)
-        }
-        return result
-    }
-    #endif
-}
-
-// MARK: Responder
-extension StaticRoute {
     /// The `RouteResponderProtocol` responder for this route.
     /// 
     /// - Parameters:
@@ -239,7 +92,7 @@ extension StaticRoute {
     ///   - function: `FunctionCallExprSyntax` that represents this route.
     ///   - middleware: Static middleware that this route will handle.
     #if StaticMiddleware
-    public func responder(
+    public mutating func responder(
         context: some MacroExpansionContext,
         function: FunctionCallExprSyntax,
         middleware: [StaticMiddleware]
@@ -247,7 +100,7 @@ extension StaticRoute {
         return response(context: context, function: function, middleware: middleware).string(escapeLineBreak: true)
     }
     #else
-    public func responder(
+    public mutating func responder(
         context: some MacroExpansionContext,
         function: FunctionCallExprSyntax
     ) throws(DestinyError) -> String? {
